Network.AWS.IoTDataPlane

Description

AWS IoT

AWS IoT-Data enables secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. It implements a broker for applications and things to publish messages over HTTP (Publish) and retrieve, update, and delete thing shadows. A thing shadow is a persistent representation of your things and their state in the AWS cloud.

Service Configuration

ioTDataPlane :: Service

API version 2015-05-28 of the Amazon IoT Data Plane SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTDataPlane.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The specified version does not match the version of the document.

RequestEntityTooLargeException

_RequestEntityTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The payload exceeds the maximum size allowed.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

MethodNotAllowedException

_MethodNotAllowedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified combination of HTTP verb and URI is not supported.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

UnauthorizedException

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform this operation.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

UnsupportedDocumentEncodingException

_UnsupportedDocumentEncodingException :: AsError a => Getting (First ServiceError) a ServiceError

The document encoding is not supported.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetThingShadow

module Network.AWS.IoTDataPlane.GetThingShadow

DeleteThingShadow

module Network.AWS.IoTDataPlane.DeleteThingShadow

UpdateThingShadow

module Network.AWS.IoTDataPlane.UpdateThingShadow

Publish

module Network.AWS.IoTDataPlane.Publish

Types