Network.AWS.IoTDataPlane.DeleteThingShadow

Description

Deletes the thing shadow for the specified thing.

For more information, see DeleteThingShadow in the AWS IoT Developer Guide .

Creating a Request

deleteThingShadow

Arguments

:: Text
dtsThingName

-> DeleteThingShadow

Creates a value of DeleteThingShadow with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtsThingName - The name of the thing.

data DeleteThingShadow

The input for the DeleteThingShadow operation.

See: deleteThingShadow smart constructor.

Request Lenses

dtsThingName :: Lens' DeleteThingShadow Text

The name of the thing.

Destructuring the Response

deleteThingShadowResponse

Arguments

:: Int
dtsrsResponseStatus

-> HashMap Text Value
dtsrsPayload

-> DeleteThingShadowResponse

Creates a value of DeleteThingShadowResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtsrsResponseStatus - -- | The response status code.

dtsrsPayload - The state information, in JSON format.

data DeleteThingShadowResponse

The output from the DeleteThingShadow operation.

See: deleteThingShadowResponse smart constructor.

Response Lenses

dtsrsResponseStatus :: Lens' DeleteThingShadowResponse Int

- | The response status code.

dtsrsPayload :: Lens' DeleteThingShadowResponse (HashMap Text Value)

The state information, in JSON format.