Network.AWS.IoTDataPlane.Publish

Description

Publishes state information.

For more information, see HTTP Protocol in the AWS IoT Developer Guide .

Creating a Request

publish

Arguments

:: Text
pTopic

-> Publish

Creates a value of Publish with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pPayload - The state information, in JSON format.

pQos - The Quality of Service (QoS) level.

pTopic - The name of the MQTT topic.

data Publish

The input for the Publish operation.

See: publish smart constructor.

Request Lenses

pPayload :: Lens' Publish (Maybe (HashMap Text Value))

The state information, in JSON format.

pQos :: Lens' Publish (Maybe Natural)

The Quality of Service (QoS) level.

pTopic :: Lens' Publish Text

The name of the MQTT topic.

Destructuring the Response

publishResponse :: PublishResponse

Creates a value of PublishResponse with the minimum fields required to make a request.

data PublishResponse

See: publishResponse smart constructor.